# rest/server/router.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	_ "embed"

	"net/http"

	"github.com/go-chi/chi/v5"
)

//go:embed openapi.yaml
var openapiYAML []byte

func mountRoutes(r chi.Router, h *handlers) {
	r.Route("/v1", func(r chi.Router) {

		r.Get("/openapi.yaml", func(w http.ResponseWriter, r *http.Request) {
			w.Header().Set("Content-Type", "text/yaml")
			_, _ = w.Write(openapiYAML)
		})

		r.Route("/taquerias", func(r chi.Router) {
			r.Get("/", h.ListTaqueria)
			r.Post("/", h.CreateTaqueria)

			r.Route("/{id}", func(r chi.Router) {
				r.Get("/", h.GetTaqueria)
				r.Patch("/", h.UpdateTaqueria)
				r.Delete("/", h.DeleteTaqueria)
			})

		})

		r.Route("/tacos", func(r chi.Router) {
			r.Get("/", h.ListTaco)
			r.Post("/", h.CreateTaco)

			r.Route("/{id}", func(r chi.Router) {
				r.Get("/", h.GetTaco)
				r.Patch("/", h.UpdateTaco)
				r.Delete("/", h.DeleteTaco)
			})

		})

		r.Route("/visitas", func(r chi.Router) {
			r.Get("/", h.ListVisita)
			r.Post("/", h.CreateVisita)

			r.Route("/{id}", func(r chi.Router) {
				r.Get("/", h.GetVisita)
				r.Patch("/", h.UpdateVisita)
				r.Delete("/", h.DeleteVisita)
			})

		})

	})

	r.Get("/healthz", func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(http.StatusOK)
		_, _ = w.Write([]byte(`{"status":"OK"}`))
	})

	// Serve the info page at the root by forwarding to the default mux, where
	// info.Register mounted it.
	r.Handle("/", http.DefaultServeMux)

}